% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{add}
\alias{add}
\title{Add together two numbers.}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la suma de dos numeros reales.
}
\examples{
add(1, 1)
add(10, 1)
}
