% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_calculation.R
\name{calculate_FIM}
\alias{calculate_FIM}
\title{Calculation of Fisher Information Matrix}
\usage{
calculate_FIM(sensitivities, times)
}
\arguments{
\item{sensitivities}{data.frame of class \code{sensFun} as returned by
\code{\link{sensitivity_inactivation}}.}

\item{times}{Numeric vector of time points where observations will be taken.}
}
\value{
Matrix with the estimation of the Fisher Information Matrix.
}
\description{
The sensitivities at the different \code{times} are calculated by 
linear interpolation of the results provided in \code{sensitivities}.
}
