% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prueba-package.R
\docType{package}
\name{CatastRoEus-package}
\alias{CatastRoEus}
\alias{CatastRoEus-package}
\title{CatastRoEus: Interface to the 'INSPIRE' services of 'Catastro de Navarra'}
\description{
Access public spatial data available under the 'INSPIRE' directive. Tools for downloading references, buildings and addresses of properties on Navarre (Spain).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/CatastRoNav/}
  \item \url{https://github.com/rOpenSpain/CatastRoNav}
  \item Report bugs at \url{https://github.com/rOpenSpain/CatastRoNav/issues}
}

}
\author{
\strong{Maintainer}: Ant Sali \email{antonio.sali@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

}
\keyword{internal}
