% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_arab.R
\name{catreus_arab_wfs_get_buildings_bbox}
\alias{catreus_arab_wfs_get_buildings_bbox}
\title{Download buildings of Araba in spatial format}
\source{
\href{https://geo.araba.eus/es/servicios-web}{GeoAraba}
}
\usage{
catreus_arab_wfs_get_buildings_bbox(x, srs, verbose = FALSE, count = NULL)
}
\arguments{
\item{x}{See \strong{Details}. It could be:
\itemize{
\item A numeric vector of length 4 with the coordinates that defines
the bounding box: \code{c(xmin, ymin, xmax, ymax)}.
\item A \code{sf/sfc} object, as provided by the \CRANpkg{sf} package.
}}

\item{srs}{SRS/CRS to use on the query. See \strong{Details}.}

\item{verbose}{Logical, displays information. Useful for debugging, default
is \code{FALSE}.}

\item{count}{integer, indicating the maximum number of features to return.
The default value \code{NULL} does not pass this parameter to the query,
and the maximum number of features would be determined by the default value
of the API service (5,000 in this case).}
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
Get the spatial data of buildings by bounding box.
}
\examples{
\donttest{
x <- c(539226.596, 4744012.338, 539236.286, 4744133.782)
bu <- catreus_arab_wfs_get_buildings_bbox(x,25830,count=50)

library(ggplot2)

ggplot(bu) +
  geom_sf()
}

}
\seealso{
\code{\link[sf:st_bbox]{sf::st_bbox()}}

\code{\link[CatastRo:catr_wfs_get_buildings]{CatastRo::catr_wfs_get_buildings_bbox()}}

\code{\link[CatastRoNav:catrnav_wfs_get_buildings_bbox]{CatastRoNav::catrnav_wfs_get_buildings_bbox()}}
}
