% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom_bizk_bu.R
\name{catreus_bizk_atom_get_buildings_db_all}
\alias{catreus_bizk_atom_get_buildings_db_all}
\title{ATOM INSPIRE: Reference database for ATOM Buildings in Bizkaia}
\source{
\href{https://www.bizkaia.eus/es/inspirebizkaia}{INSPIREBIZKAIA}
}
\usage{
catreus_bizk_atom_get_buildings_db_all(
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cache}{A logical whether to do caching. Default is \code{TRUE}. See
\strong{About caching} section on \code{\link[=catreus_set_cache_dir]{catreus_set_cache_dir()}}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source file.}

\item{cache_dir}{A path to a cache directory. On missing value the function would store the cached files on a temporary dir (See \code{\link[base:tempfile]{base::tempdir()}}).}

\item{verbose}{Logical, displays information. Useful for debugging, default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble]{tibble} with the information requested:
\itemize{
\item \code{munic}: Name of the municipality.
\item \code{url}: url for downloading information of the corresponding municipality.
\item \code{date}: Reference date of the data.
}
}
\description{
Create a database containing the urls provided in the INSPIRE ATOM service
for extracting Buildings.
}
\examples{
\donttest{
catreus_bizk_atom_get_buildings_db_all()
}
}
\seealso{
Other ATOM: 
\code{\link{catreus_bizk_atom_get_addresses}()},
\code{\link{catreus_bizk_atom_get_addresses_db_all}()},
\code{\link{catreus_bizk_atom_get_buildings}()},
\code{\link{catreus_bizk_atom_get_parcels}()},
\code{\link{catreus_bizk_atom_get_parcels_db_all}()}

Other buildings: 
\code{\link{catreus_bizk_atom_get_buildings}()}
}
\concept{ATOM}
\concept{buildings}
