% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_bizk.R
\name{catreus_bizk_wfs_get_parcels_bbox}
\alias{catreus_bizk_wfs_get_parcels_bbox}
\title{Download cadastral parcels of Bizkaia in spatial format}
\source{
\href{https://www.bizkaia.eus/es/inspirebizkaia#wfs}{INSPIREBIZKAIA}
}
\usage{
catreus_bizk_wfs_get_parcels_bbox(x, srs, verbose = FALSE, count = NULL)
}
\arguments{
\item{x}{See \strong{Details}. It could be:
\itemize{
\item A numeric vector of length 4 with the coordinates that defines
the bounding box: \code{c(xmin, ymin, xmax, ymax)}.
\item A \code{sf/sfc} object, as provided by the \CRANpkg{sf} package.
}}

\item{srs}{SRS/CRS to use on the query. See \strong{Details}.}

\item{verbose}{Logical, displays information. Useful for debugging, default
is \code{FALSE}.}

\item{count}{integer, indicating the maximum number of features to return.
The default value \code{NULL} does not pass this parameter to the query,
and the maximum number of features would be determined by the default value
of the API service (5,000 in this case).}
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
Get the spatial data of cadastral parcels by bounding box.
}
\examples{
\donttest{
x <- c(43.312, -2.994, 43.313, -2.993)

cp <- catreus_bizk_wfs_get_parcels_bbox(x, srs = 25830)

library(ggplot2)

ggplot(cp) +
  geom_sf()
}

}
\seealso{
\code{\link[sf:st_bbox]{sf::st_bbox()}}

\code{\link[CatastRo:catr_wfs_get_parcels]{CatastRo::catr_wfs_get_parcels_bbox()}}

\code{\link[CatastRoNav:catrnav_wfs_get_parcels_bbox]{CatastRoNav::catrnav_wfs_get_parcels_bbox()}}

Other parcels: 
\code{\link{catreus_arab_wfs_get_parcels_bbox}()},
\code{\link{catreus_bizk_atom_get_parcels}()},
\code{\link{catreus_bizk_atom_get_parcels_db_all}()},
\code{\link{catreus_gipu_wfs_get_parcels_bbox}()}
}
\concept{parcels}
