% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_gipu.R
\name{catreus_gipu_wfs_get_address_bbox}
\alias{catreus_gipu_wfs_get_address_bbox}
\title{Download addresses of Gipuzkoa in spatial format}
\source{
\href{https://b5m.gipuzkoa.eus/web5000/en/inspire-services}{Gipuzkoa Spatial Data Infrastructure}
}
\usage{
catreus_gipu_wfs_get_address_bbox(x, srs, verbose = FALSE, count = NULL)
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
Get the spatial data of addresses by bounding box.
}
\examples{
\donttest{

x <- c(-1.979990, 43.312830, -1.977990, 43.314053)
x <- c(-2.517969, 43.023601, -2.384621, 43.313489)
ad <- catreus_gipu_wfs_get_address_bbox(x, srs=4326, count=50)

library(ggplot2)

ggplot(ad) +
  geom_sf()
}

}
\seealso{
\code{\link[sf:st_bbox]{sf::st_bbox()}}

\code{\link[CatastRo:catr_wfs_get_address]{CatastRo::catr_wfs_get_address_bbox()}}

\code{\link[CatastRoNav:catrnav_wfs_get_address_bbox]{CatastRoNav::catrnav_wfs_get_address_bbox()}}
}
