% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_bizk.R
\name{catrneus_bizk_wfs_get_address_bbox}
\alias{catrneus_bizk_wfs_get_address_bbox}
\title{Download addresses of Bizkaia in spatial format}
\source{
\href{https://www.bizkaia.eus/es/inspirebizkaia#wfs}{INSPIREBIZKAIA}
}
\usage{
catrneus_bizk_wfs_get_address_bbox(x, srs, verbose = FALSE, count = NULL)
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
Get the spatial data of addresses by bounding box.
}
\examples{
\donttest{
x <- c(43.312, -2.994, 43.313, -2.993)

ad <- catrneus_bizk_wfs_get_address_bbox(x, srs = 25830)

library(ggplot2)

ggplot(ad) +
  geom_sf()
}

}
\seealso{
\code{\link[sf:st_bbox]{sf::st_bbox()}}

\code{\link[CatastRo:catr_wfs_get_address]{CatastRo::catr_wfs_get_address_bbox()}}

\code{\link[CatastRoNav:catrnav_wfs_get_address_bbox]{CatastRoNav::catrnav_wfs_get_address_bbox()}}
}
