% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_calculation.R
\name{criterium_modE}
\alias{criterium_modE}
\title{Modified-E Optimality Criterium}
\usage{
criterium_modE(FIM, eig_tol = 1e-10)
}
\arguments{
\item{FIM}{Matrix with the values of the Fisher Information Matrix}

\item{eig_tol}{Tolerance for the eigen values. If any eigen value is lower than this
value, the FIM is singular and a high value (1e20) is returned. 1e-10 by default.}
}
\description{
Modified-E Optimality Criterium
}
