% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_limits.R
\name{detection_bigelow}
\alias{detection_bigelow}
\title{Detection limit of the Bigelow model}
\usage{
detection_bigelow(pars, temperature, limit)
}
\description{
Detection limit of the Bigelow model
}
