% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dHYPERPO_single}
\alias{dHYPERPO_single}
\title{Function to obtain the dHYPERPO for a single value x}
\usage{
dHYPERPO_single(x, mu = 1, sigma = 1, log = FALSE)
}
\arguments{
\item{x}{numeric value for x.}

\item{mu}{numeric value for nu.}

\item{sigma}{numeric value for sigma.}

\item{log}{logical value for log.}
}
\value{
returns the pmf for a single value x.
}
\description{
Function to obtain the dHYPERPO for a single value x
}
