% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distri.R
\name{dxx}
\alias{dxx}
\alias{pxx}
\alias{qxx}
\alias{rxx}
\title{distri}
\usage{
dxx(x)

pxx(q)

qxx(p)

rxx(n)
}
\arguments{
\item{x}{es un valor de X.}

\item{q}{es un cuantil.}

\item{p}{es una probabilidad.}

\item{n}{es numero de obs.}
}
\description{
Esta funcion sirve de ejemplo para documentar distribuciones.
}
\examples{
dxx(2.3)
dxx(3.5)
}
