% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f11_cpp}
\alias{f11_cpp}
\title{Function to obtain F11 with C++.}
\usage{
f11_cpp(gamma, lambda, maxiter_series = 10000L, tol = 1e-10)
}
\arguments{
\item{gamma}{numeric value for gamma.}

\item{lambda}{numeric value for lambda.}

\item{maxiter_series}{numeric value.}

\item{tol}{numeric value.}
}
\value{
returns the F11 value.
}
\description{
Function to obtain F11 with C++.
}
