% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plotea los accidentes en un mapa de USA}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{El numero del estado de USA como usado en FARS.
Debe ser numerico entero.}

\item{year}{El anio del analisis(numerico o entero)}
}
\value{
Returna un ploteo de accidentes basado en las entradas \code{state.num} y
\code{year}. Returna un error si el estado o el anio no existen en el conjunto
de datos.
}
\description{
La funcion acepta un numero de estado y un anio y plotea los accidentes en un mapa 
simple. El numero de estado debe ser entero o numerico y debe existir
en los datos FARS, de otra forma la funcion termina con un error. Tambien
returna un error si el archivo de datos para ese anio no existe.
}
\examples{
\dontrun{
fars_map_state(45, 2014)

# Results in an error
fars_map_state(43, 2013)
fars_map_state(50, 2014)
}

}
