% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Lee un archivo CSV}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Camino al archivo CSV (character)}
}
\value{
La funcion retura un tibble (data.frame) basado en el archivo CSV.
}
\description{
La funcion lee un archivo CSV definino por \code{filename} argumento y retorna
un tibble. Si el path esta incorrecto, la funcion termina con un error.
}
\examples{
\dontrun{
accident_2015 <- fars_read("./data/accident_2013.csv.bz2")
}

}
