% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Cuenta el numero de accidentes por mes y anio}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{UN vector o lista de anios (numericos o enteros) que van a ser buscados en
los datos.}
}
\value{
Retorna unpivot tibble (data frame) con meses en las filas y los anios seleccionados
en las columnas conteniendo el numero de accidentes. Retorna un warning por cada anio de 
entrada que no exista en el datasets. Retorna un error (no
results) si un numero diferente a entero es presentado.
}
\description{
Basado en la lista de anios, la funcion calculael numero de accidentes
en USA en una base mensual. El archivo de accidentes necesita estar en el directorio 
de trabajo, el anio puede ser pasado como lista o vector.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
