% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_alias.R
\name{hola1}
\alias{hola1}
\alias{hola2}
\title{hola1}
\usage{
hola1(nombre)

hola2(nombre)
}
\arguments{
\item{nombre}{el nombre del usuario}
}
\description{
Esta funcion saluda en forma 1
}
\examples{
hola1('olga')
}
