% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sensitivity.R
\name{inactivation_sens_handler}
\alias{inactivation_sens_handler}
\title{Handler for the calculation of sensitivities of inactivation models}
\usage{
inactivation_sens_handler(model_parms, inactivation_model, times,
  temp_profile, parms_fix)
}
\arguments{
\item{model_parms}{A named vector or list with the values of the model parameters. See
the documentation of bioinactivation::predict_inactivation.}

\item{inactivation_model}{A character defining the inactivation model to use. See
the documentation of bioinactivation::predict_inactivation.}

\item{times}{A numeric vector describing the points where the solution will be calculated.
See the documentation of bioinactivation::predict_inactivation.}

\item{temp_profile}{A data frame describing the temperature profile. See
the documentation of bioinactivation::predict_inactivation.}

\item{parms_fix}{A named vector or list with the values of the known model parameters. See
the documentation of bioinactivation::predict_inactivation.}
}
\description{
Handler for the calculation of sensitivities of inactivation models
}
