% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{infoBoxPROMiDAT}
\alias{infoBoxPROMiDAT}
\title{Information box.}
\usage{
infoBoxPROMiDAT(titulo, valor, icono)
}
\arguments{
\item{titulo}{Title text.}

\item{valor}{The value to display in the box. Usually a number or short text.}

\item{icono}{An icon tag, created by icon.}
}
\value{
An HTML element.
}
\description{
Information box.
}
\examples{
library(shiny)
infoBoxPROMiDAT("Title", "Value", icon("info"))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
