% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OED_isothermal.R
\name{isothermal_OED}
\alias{isothermal_OED}
\title{Optimal Experiment Design of isothermal inactivation}
\usage{
isothermal_OED(model, pars, n_points, min_time, max_time, min_temp,
  max_temp, opts = NULL)
}
\arguments{
\item{model}{character string defining the inactivation model to use.}

\item{pars}{list defining the nominal model parameters.}

\item{n_points}{numerical stating the number of data points.}

\item{min_time}{numerical stating the lower limit for the time points.}

\item{max_time}{numerical stating the upper limit for the time points.}

\item{min_temp}{numerical stating the lower limit for the temperature.}

\item{max_temp}{numerical stating the upper limit for the temperature.}

\item{opts}{options for the MEIGO algorithm. By default, a maximum of 2000
function evaluations with local finish with the DHC algorithm 
(see help from MEIGO).}
}
\value{
A MEIGO object
}
\description{
OED of microbial inactivation experiments.
}
\examples{
pars <- list(temp_crit = 55,
n = 1.5,
k_b = 0.1)
OED <- isothermal_OED("Peleg", pars,
                            n_points=10, min_time=0, max_time=100, min_temp=52, max_temp=60,
                            opts = NULL)
OED$optim$xbest

}
