% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OED_isothermal_limit.R
\name{isothermal_OED_limit}
\alias{isothermal_OED_limit}
\title{OED of isothermal microbial inactivation with detection limit}
\usage{
isothermal_OED_limit(model, pars, limit, n_points, min_time, max_time,
  min_temp, max_temp, criterium, opts = NULL)
}
\arguments{
\item{model}{character string defining the inactivation model to use.}

\item{pars}{list defining the nominal model parameters.}

\item{limit}{numerical value describing the maximum number of log-reductions
that can be identified in the experiment limit = logDL - logN0, where DL
is the detection limit.}

\item{n_points}{numerical stating the number of data points.}

\item{min_time}{numerical stating the lower limit for the time points.}

\item{max_time}{numerical stating the upper limit for the time points.}

\item{min_temp}{numerical stating the lower limit for the temperature.}

\item{max_temp}{numerical stating the upper limit for the temperature.}

\item{criterium}{character string defining the criterium to use.}

\item{opts}{options for the MEIGO algorithm. By default, a maximum of 2000
function evaluations with local finish with the DHC algorithm 
(see help from MEIGO).}
}
\value{
A MEIGO object
}
\description{
OED of isothermal microbial inactivation with detection limit
}
\examples{
pars <- list(temp_crit = 55,
n = 1.5,
k_b = 0.1)
OED <- isothermal_OED_limit("Peleg", pars, limit=7,
                            n_points=10, min_time=0, max_time=100, min_temp=52, max_temp=60, criterium="D",
                            opts = NULL)
OED$optim$xbest

}
