% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{labelInput}
\alias{labelInput}
\title{Create a label that can be used to show text.}
\usage{
labelInput(inputId, value = "")
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{value}{Initial value.}
}
\value{
An HTML element.
}
\description{
Create a label that can be used to show text.
}
\examples{
labelInput("id", "data")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
