% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{labels_loadeR}
\alias{labels_loadeR}
\title{Returns a vector of keys to translate with tr.}
\usage{
labels_loadeR()
}
\value{
a vector of keys.
}
\description{
Returns a vector of keys to translate with tr.
}
\examples{
labels_loadeR()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
