% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadeR.R
\docType{package}
\name{loadeR}
\alias{loadeR}
\title{Read Data System}
\description{
Provides a framework to load text and excel files through a 'shiny' graphical interface. It allows renaming, transforming, ordering and removing variables. It includes basic exploratory methods such as the mean, median, mode, normality test, histogtram and correlation.
}
\details{
\tabular{ll}{
Package: \tab loadeR\cr
Type: \tab Package\cr
Version: \tab 1.1.7\cr
Date: \tab 2023-07-19\cr
License: \tab GPL (>=2)\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://promidat.website}
  \item \url{https://github.com/PROMiDAT/loadeR}
  \item Report bugs at \url{https://github.com/PROMiDAT/loadeR/issues}
}

}
\author{
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>\cr
\itemize{
  \item Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
  \item Diego Jiménez Alvarado
  \item Joseline Quirós Mendez
}
}
\keyword{package}
