% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Crea un filename}
\usage{
make_filename(year)
}
\arguments{
\item{año}{Numerico o entero que indica el año del set de datos.}
}
\value{
Returna un string de caracteres en el formato "accident_<year>.csv.bz2" que puede 
ser usado como un nombre de archivo
}
\description{
La funcion creea un filename para el archivo .csv.bz2 basado en el \code{year}
argumento con la forma "accident_<year>.csv.bz2". Requires un valor de entrada numerico o
entero, de otra forma finaliza con un error.
}
\examples{
\dontrun{
makefilename(2016)
}
}
