% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{media}
\usage{
media(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mean of a vector.
}
\description{
Esta funcion sirve para obtener la media de un vector.
}
\examples{
x <- 1:7
media(x)
mean(x)
}
