% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{menu.idioma}
\alias{menu.idioma}
\title{HTML for language menu.}
\usage{
menu.idioma()
}
\value{
An HTML element.
}
\description{
HTML for language menu.
}
\examples{
menu.idioma()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
