% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_max.R
\name{mini_max}
\alias{mini_max}
\title{minimo y maximo}
\usage{
mini_max(n, media, desvi)
}
\arguments{
\item{n}{tamano muestral}

\item{media}{media poblacional}

\item{desvi}{desviacion estadar}
}
\value{
un vector con el minimo y el maximo
}
\description{
minimo y maximo
}
\examples{

mini_max(10,20,4)
mini_max(50,25,3)

}
