% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misdatos.R
\docType{data}
\name{misdatos}
\alias{misdatos}
\title{Mis datos de prueba}
\format{
This data frame with 3 rows and 2 variables:
\describe{
\item{estatura}{altura de las personas en centimetros}
\item{peso}{peso corporal de las personas en kilogramos}
}
}
\usage{
misdatos
}
\description{
Estos son uno datos de prueba para mostrarle a Deisy como se crean.
}
\keyword{datasets}
