% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos.R
\name{mod_carga_datos_server}
\alias{mod_carga_datos_server}
\title{carga_datos Server Functions.}
\usage{
mod_carga_datos_server(
  id,
  updateData,
  modelos,
  codedioma,
  paquete = "predictoR"
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{modelos}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}

\item{paquete}{indicates if the data is going to be used for exploratory, predictive, or regression analysis.}
}
\value{
shiny server module.
}
\description{
carga_datos Server Functions.
}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
