% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos.R
\name{mod_carga_datos_ui}
\alias{mod_carga_datos_ui}
\title{carga_datos UI Function}
\usage{
mod_carga_datos_ui(id, title, paquete = "predictoR")
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{title}{Display title for tab.}

\item{paquete}{indicates if the data is going to be used for exploratory, predictive, or regression analysis.}
}
\value{
shiny ui module.
}
\description{
carga_datos UI Function
}
\author{
Joseline Quiros <joseline.quiros@promidat.com>
}
