% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_correlacion.R
\name{mod_correlacion_server}
\alias{mod_correlacion_server}
\title{correlacion Server Function}
\usage{
mod_correlacion_server(id, updateData, codedioma)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}
}
\value{
shiny server module.
}
\description{
correlacion Server Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
