% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_correlacion.R
\name{mod_correlacion_ui}
\alias{mod_correlacion_ui}
\title{correlacion UI Function}
\usage{
mod_correlacion_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui module.
}
\description{
correlacion UI Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
