% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dispersion.R
\name{mod_dispersion_server}
\alias{mod_dispersion_server}
\title{dispersion Server Function}
\usage{
mod_dispersion_server(id, updateData, codedioma)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}
}
\value{
shiny server module.
}
\description{
dispersion Server Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
