% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dispersion.R
\name{mod_dispersion_ui}
\alias{mod_dispersion_ui}
\title{dispersion UI Function}
\usage{
mod_dispersion_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui module.
}
\description{
dispersion UI Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
