% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones.R
\name{mod_distribuciones_server}
\alias{mod_distribuciones_server}
\title{distribuciones Server Function}
\usage{
mod_distribuciones_server(id, updateData, codedioma)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}
}
\value{
shiny server module.
}
\description{
distribuciones Server Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
