% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_distribuciones.R
\name{mod_distribuciones_ui}
\alias{mod_distribuciones_ui}
\title{distribuciones UI Function}
\usage{
mod_distribuciones_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui module.
}
\description{
distribuciones UI Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
