% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal.R
\name{mod_normal_server}
\alias{mod_normal_server}
\title{normal Server Function}
\usage{
mod_normal_server(id, updateData, codedioma)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}
}
\value{
shiny server module.
}
\description{
normal Server Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
