% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normal.R
\name{mod_normal_ui}
\alias{mod_normal_ui}
\title{normal UI Function}
\usage{
mod_normal_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui module.
}
\description{
normal UI Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
