% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_numerico.R
\name{mod_r_numerico_server}
\alias{mod_r_numerico_server}
\title{r_numerico Server Function}
\usage{
mod_r_numerico_server(id, updateData, codedioma)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{updateData}{shiny reactive values.}

\item{codedioma}{shiny reactive values.}
}
\value{
shiny server module.
}
\description{
r_numerico Server Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
