% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_r_numerico.R
\name{mod_r_numerico_ui}
\alias{mod_r_numerico_ui}
\title{r_numerico UI Function}
\usage{
mod_r_numerico_ui(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
shiny ui module.
}
\description{
r_numerico UI Function
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
