% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidada2.R
\name{multi2}
\alias{multi2}
\title{multi2}
\usage{
multi2(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la multiplicacion de dos numeros reales.
}
\examples{
multi2(1, 1)
multi2(10, 1)
}
