% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OED_noPenalty.R
\name{objective_D}
\alias{objective_D}
\title{Objective Function for the D Criterium}
\usage{
objective_D(times, sensitivities)
}
\arguments{
\item{times}{A numeric vector of points where the FIM will be calculated.}

\item{sensitivities}{An object returned by sensitivity_inactivation.}
}
\description{
Objective Function for the D Criterium
}
