% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OED_penalty.R
\name{objective_D_penalty}
\alias{objective_D_penalty}
\title{Objective Function for the D Criterium with Penalty}
\usage{
objective_D_penalty(times, sensitivities, time_min, ...)
}
\arguments{
\item{times}{Numeric vector of points where the FIM is calculated.}

\item{sensitivities}{An object returned by sensitivity_inactivation.}

\item{time_min}{Numeric defining the minimum time between measurements.}

\item{...}{Additional arguments passed to penalty_function.}
}
\description{
Objective Function for the D Criterium with Penalty
}
