% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OED_noPenalty.R
\name{objective_Emod}
\alias{objective_Emod}
\title{Objective Function for the modified-E Criterium}
\usage{
objective_Emod(times, sensitivities)
}
\arguments{
\item{times}{A numeric vector of points where the FIM will be calculated.}

\item{sensitivities}{An object returned by sensitivity_inactivation.}
}
\description{
Objective Function for the modified-E Criterium
}
