% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_S3.R
\name{plot.IsoSensitivities}
\alias{plot.IsoSensitivities}
\title{Plotting of IsoSensitivities objects}
\usage{
\method{plot}{IsoSensitivities}(x, y = NULL, ..., limit = NULL)
}
\description{
Plotting of IsoSensitivities objects
}
