% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_S3.R
\name{plot.OEDinactivation}
\alias{plot.OEDinactivation}
\title{Plot of OEDinactivation}
\usage{
\method{plot}{OEDinactivation}(x, y = NULL, ...)
}
\arguments{
\item{x}{An instance of OEDinactivation}

\item{y}{Ignored}

\item{...}{Ignored}
}
\description{
Plot of OEDinactivation
}
