% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_S3.R
\name{plot.parCorrelation}
\alias{plot.parCorrelation}
\title{Correlation Plot of Parameter Sensitivities}
\usage{
\method{plot}{parCorrelation}(x, y = NULL, ...)
}
\arguments{
\item{x}{Instance of parCorrelation}

\item{y}{Ignored}

\item{...}{Ignored}
}
\description{
Makes a correlation plot of the sensitivities between model
parameters.
}
