% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{plot.sumita}
\alias{plot.sumita}
\title{Plot components from ETS model}
\usage{
\method{plot}{sumita}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{sumita}.}

\item{...}{Other plotting parameters to affect the plot.}

\item{object}{Object of class \dQuote{sumita}.}

\item{range.bars}{Logical indicating if each plot should have a bar at its
right side representing relative size. If NULL, automatic selection takes
place.}
}
\value{
None. Function produces a plot
}
\description{
Produces a plot of the level, slope and seasonal components from an ETS
model.
}
\details{
\code{plot} will produce an equivalent plot as a ggplot object.
}
\examples{

fit <- suma(10, 5)
plot(fit)

}
\seealso{
\code{\link{suma}}
}
\author{
Rob J Hyndman & Mitchell O'Hara-Wild
}
\keyword{hplot}
