% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_ref_temp.R
\name{refTemp_optim_handler}
\alias{refTemp_optim_handler}
\title{Hanlder for the Optimization of Reference Temperature}
\usage{
refTemp_optim_handler(temp_ref, inactivation_model, parms, temp_profile,
  parms_fix, n_times, temp_ref0)
}
\arguments{
\item{temp_ref}{New value of the reference temperature.}

\item{inactivation_model}{Character identifying the inactivation model
to use for the calculation.}

\item{parms}{Numeric vector with the nominal values of the model parameters.}

\item{temp_profile}{Data frame describing the environmental conditions.}

\item{parms_fix}{Nominal value of the parameters not considered for the
sensitivity.}

\item{n_times}{Numeric value specifying the nombers of time points where 
the sensitivity functions will be calculated. 100 by default.}

\item{temp_ref0}{Initial value of the reference temperature.}
}
\description{
Hanlder for the Optimization of Reference Temperature
}
