% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(paquete = "predictoR", ...)
}
\arguments{
\item{paquete}{indicates if the data is going to be used for exploratory, predictive, or regression analysis.}

\item{...}{A series of options to be used inside the app.}
}
\value{
No return value, run a shiny application.
}
\description{
Run the Shiny Application
}
