% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_sensitivities.R
\name{sensitivities_Bigelow}
\alias{sensitivities_Bigelow}
\title{Local sensitivities of the Bigelow model}
\usage{
sensitivities_Bigelow(exp_design, pars)
}
\arguments{
\item{exp_design}{data.frame with two columns named \code{times} and 
\code{temperature} describing the experiment design.}

\item{pars}{list defining the model parameters according to the rules defined
in the \code{bioinactivation} package.}
}
\value{
A data frame with the same number of rows as \code{exp_design} with 
additional column for local sensitivities. These are named D_R and z for local
sensitivities and D_R_scaled and z_scaled for scaled locan sensitivities.
}
\description{
Local sensitivities of the Bigelow model
}
