% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_sensitivities.R
\name{sensitivities_Mafart}
\alias{sensitivities_Mafart}
\title{Local sensitivities of the Mafart model}
\usage{
sensitivities_Mafart(exp_design, pars)
}
\arguments{
\item{exp_design}{data.frame with two columns named \code{times} and 
\code{temperature} describing the experiment design.}

\item{pars}{list defining the model parameters according to the rules defined
in the \code{bioinactivation} package.}
}
\description{
Local sensitivities of the Mafart model
}
