% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hp.R
\name{simulate_hp}
\alias{simulate_hp}
\title{The simulate_hp}
\usage{
simulate_hp(sigma, mu)
}
\arguments{
\item{sigma}{a single value of the sigma parameter.}

\item{mu}{a single value of the mu parameter.}
}
\value{
a single value for the HYPERPO distribution.
}
\description{
Auxiliar function to generate a single observation for HYPERPO.
}
