% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_tres.R
\name{suma_tres}
\alias{suma_tres}
\alias{multi_tres}
\title{Title}
\usage{
suma_tres(a, b, c)

multi_tres(a, b, c)
}
\arguments{
\item{a}{primer valor.}

\item{b}{segundo valor.}

\item{c}{tercer valor.}
}
\value{
un resultado
}
\description{
Title
}
\examples{
suma_tres(a=3, b=5, c=9)
multi_tres(a=6, b=2, c=8)
}
