% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{tabBoxPrmdt}
\alias{tabBoxPrmdt}
\title{Tabset panel with options menu.}
\usage{
tabBoxPrmdt(..., id = NULL, title = NULL, opciones = NULL, open = NULL)
}
\arguments{
\item{...}{tabPanel() elements to include in the tabset.}

\item{id}{If provided, you can use input$id in your server logic to determine which of the current tabs is active. The value will correspond to the value argument that is passed to tabPanel().}

\item{title}{Text or input to add on the opposite side of the tabs.}

\item{opciones}{list of html options to add on the footer of the tabset.}

\item{open}{Class to assign first option, for example to start open.}
}
\value{
An HTML element.
}
\description{
Tabset panel with options menu.
}
\examples{
library(shiny)
tabBoxPrmdt(id = "id", title = "title", tabPanel("Tab1"))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
