% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{tabsOptions}
\alias{tabsOptions}
\title{Options menu in footer for tabBoxPrmdt (tabsetPanel).}
\usage{
tabsOptions(
  botones = list(paste(labelInput("opts"), icon("gear"))),
  widths = 100,
  heights = 50,
  tabs.content = list(""),
  id = NULL
)
}
\arguments{
\item{botones}{list of icons to each option of the menu. Minimum 1, maximum 5.}

\item{widths}{vector of widths to each option of the menu. Minimum 1, maximum 5.}

\item{heights}{vector of heights to each option of the menu. Minimum 1, maximum 5.}

\item{tabs.content}{list of UI elements to include within each menu option. Minimum 1, maximum 5.}

\item{id}{If provided, you can use input$id in your server logic to get the element.}
}
\value{
An HTML element.
}
\description{
Options menu in footer for tabBoxPrmdt (tabsetPanel).
}
\examples{
tabsOptions()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
