% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{tr}
\alias{tr}
\title{Returns a translate text (user defined).}
\usage{
tr(text, idioma = "es")
}
\arguments{
\item{text}{text to translate.}

\item{idioma}{language to use. For example: "en".}
}
\value{
a translate text.
}
\description{
Returns a translate text (user defined).
}
\examples{
tr("data", "en")

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
