% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{translation.loadeR}
\alias{translation.loadeR}
\title{Returns a list of sentences with their translation in different languages.}
\usage{
translation.loadeR()
}
\value{
a list of sentences with their translation in different languages.
}
\description{
Returns a list of sentences with their translation in different languages.
}
\examples{
translation.loadeR()

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
