% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{updateLabelInput}
\alias{updateLabelInput}
\title{Change the value of a label input on the client.}
\usage{
updateLabelInput(session, labelid, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer. Default is getDefaultReactiveDomain().}

\item{labelid}{The id of the input object.}

\item{value}{Initial value.}
}
\value{
An HTML element.
}
\description{
Change the value of a label input on the client.
}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
