% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianza.R
\name{varianza}
\alias{varianza}
\title{varianza}
\usage{
varianza(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The variance of a vector.
}
\description{
Esta funcion sirve para obtener la varianza de un vector.
}
\examples{
x <- 1:7
varianza(x)
var(x)
}
