% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misdatos.R
\docType{data}
\name{misdatos}
\alias{misdatos}
\title{Mis datos de prueba}
\format{Time series data}
\source{
Australian Bureau of Statistics.
}
\usage{
misdatos
}
\arguments{
\item{estatura}{altura de las personas en centimetros.}

\item{peso}{peso corporal de las personas en kilogramos.}
}
\description{
Estos son uno datos de prueba para mostrarle a Deisy como se crean.
}
\keyword{datasets}
