% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidada.R
\name{multi}
\alias{multi}
\alias{myaux}
\title{multi}
\usage{
multi(x, y)

myaux(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la multiplicacion de dos numeros reales.
}
\examples{
multi(1, 1)
multi(10, 1)
asbio::shade.norm(x=1.2,sigma=1,mu=0,tail="lower")
}
